backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
google-genai>=1.16
sse-starlette>=2.1
python-dotenv>=1.0
pydantic>=2.10

# test
pytest>=8.3
pytest-asyncio>=0.24
httpx>=0.27
